requirements.txt
click
filemagic
attr
colorama
